SUMMARY: Adding general framework-level support for TPLs.
BUGS: 4289

Here I have added framework-level support for TPLs.  Now, BLAS,
LAPACK, Boost, and even MPI are all handled in a consistent automated
way.  Each PACAKGE_DIR/cmake/Dependencies.cmake file now lists for
additional lists for TPLs (i.e. LIB/TEST and OPTIONAL/REQUIRED).  Even
Dependencies.cmake file has to now list 8 different lists.  This is
needed in order to catch mispelled varaibles.

One important change represented here is that the option
Trilinos_ENABLE_MPI is no more and is replaced with TPL_ENABLE_MPI.
This is to be consistent with all of the other TPLs.  I could have
used Trilinos_ENABLE_${TPL} instead of TPL_ENABLE_${TPL} but I wanted
to make it clear that BLAS, LAPACK, Boost, MPI, etc. are external TPLs
and not native Trilinos packages and I thing this does that.

I will document the system a little later after I get some more TPL
support in like for Zoltan.  Now we can go ahead and finish up the
CMakeification of Zoltan.

I also have yet to put in unit tests for this new infrastructure but
that will come next.

One other small change that I have to mention:

(*) Epetra_C_wrappers.cpp: I had to add an include for
Epetra_ConfigDefs.h at the very top of the file in order for
EPETRA_MPI to be correctly defined.  The CMake system (and I suspect
also the autotools system) was hacked to define EPETRA_MPI on the
command line.  This is a big no no and I can't believe that someone
did not already catch this.  We decided a long time ago that we were
not going to support the no config.h file mode.  I also removed
worthless ifdefs for __cplusplus.  This is a C++ file and can never be
compiled as a C file.  There are C++ classes in this file for goodness
sake.  I am sure this was just a silly copy and paste error that has
not been corrected.

This has been tested with:
- CMake serial debug (all 149 tests run passed)
- CMake mpi opt (all 137 tests run passed)