Removing redundant standard TriBITS includes
Removing these includes speeds up the configure by a good bit and they just
clutter up these files.  Win win.

A side effect of the script that made these changes is that it adds a newline
to files that don't end in a newline.

I removed the upates to SEACAS and STK because of the problem with integration
with SIERRA.