Enable non-developer mode
Developer mode can not be switched during the configuration stage
through
```cmake
-D ForTrilinos_ENABLE_DeveloperMode=ON
```

In developer mode, we automatically generated .f90 and .cxx from C++
header files, store them in the source directory, and compile the
library.

In the non-developer mode, we only compile the files that are in the
source directory.