ForTrilinos: simplifying CMake logic
Thanks to insight in TriBITSPub/TriBITS#178, it is clear now that we
need to have
```CMake
INLCUDE_DIRECTORIES(${CMAKE_CURRENT_BINARY_DIR})
```
in `CMakeLists.txt` for very directory containing Fortran source code.
This would add the binary directory with auto-generated `.mod` file in
it to the list of directories that would be included.