ForTrilinos: fixed IoC Fortran test
To have no preconditioner one must declare "Preconditioner Type" to be
"None". Simply removing it does not work as if there are any
preconditioners enabled, the first one would be used. For the current
Docker configuration, this is Ifpack, which causes a problem as it
expects an Epetra_Operator, but instead gets a Tpetra::Operator.