Tpetra: fix release() for MultiVector
Ghe generated code for Multivector `release()` was different from all
other Tpetra classes (even from CrsMatrix which also has
a "classic" template argument) for unknow reason, and instead of calling
`delete` on RCP, it was calling `delete` only on RCP data. Thus, RCP
became dangling, which lead to problems in Trilinos Debug mode.

This commit not a fix, but just a workaround to explicitly attach
"unref" feature to the class.