simple: fix IoC operator test for MPI
It had to properly import parts of the X vector.

Second, in MPI we continuously got
  Note: The following floating-point exceptions are signalling: IEEE_DENORMAL

While the situation is still not 100% clear, some investigation seems to
indicate that the vector Y coming from Belos may be unitialized, and can
potentially contain denormalized or signaling NaN entries. This may be
the case only for $beta = 0$. The workaround is to treat that case
specially, and instead of doing $Y = 0*Y$, so $Y = 0$ first.