Fix assignment from null array
Not clear why this was working in GCC. Whenever a SWIG execption function
fails, it returns an empty/null value, which in the case of a SwigArray
is `{null, 0}`. It's unspecified behavior to do `c_f_pointer` with
zero-sized arrays. GCC was somehow working, but flang chose to convert
this result to a null fortran pointer, which then is *assigned* to `iarr`
before the throw condition is checked.

The solution is to have the test return pointers rather than automatically
allocating arrays.