Properly test the solution in the linear solver test
The only way to have robust test is to check that the relative residual
of the final solution is less than the specified tolerance. Prior to
this patch we tested that the solution was "close" to the true solution,
however that is not robust, and for example depends on the linear solver
configuration. We were lucky enough for that to happen with ILUT
solvers, but switching to Symmetric Gauss-Seidel broke that.